Morse device for CAN ESP32
