+++ J345/src/J5Device.hpp
@@ -6,11 +6,9 @@
 #include "rover_lib2/sensors/push_button.hpp"
 #include "rover_lib2/helpers/loop_timer.hpp"
 #include "config.hpp"
-#include "MorsePlayer.hpp"
 
 #include "rover_can2/rover_can2.hpp"
 #include "rover_can2/msgs/arm_joint_cmd.hpp"
-#include "rover_can2/msgs/morse_code.hpp"
 
 #include <cstring>
 
@@ -27,13 +25,8 @@ class J5Device
     static constexpr uint64_t CAN_SEND_PERIOD_MS = static_cast<uint64_t>(ROUND(1'000.0F / CAN_SEND_FREQUENCY));
     static constexpr uint64_t CAN_WATCHDOG_VALIDITY_PERIOD = static_cast<uint64_t>(1'000.0F / CAN_SEND_FREQUENCY * 2.0F);
 
-    // --- Morse code message reassembly ---
-    static constexpr size_t MORSE_MAX_LEN = 64;
-    static constexpr uint64_t MORSE_FRAME_TIMEOUT_MS = 500ULL;
-
     using JointCanDeviceT = RoverCan2::Device<RoverCan2::SubscriberMember<RoverCan2::Msgs::ArmJointCmd, J5Device>,
                                               RoverCan2::Publisher<RoverCan2::Msgs::ArmJointStatus>>;
-    using MorseCodeCanDeviceT = RoverCan2::Device<RoverCan2::SubscriberMember<RoverCan2::Msgs::MorseCode, J5Device>>;
 
   public:
     J5Device() = default;
@@ -51,15 +44,7 @@ class J5Device
             return;
         }
 
-        // Abandon a stalled partial morse message
-        if (_morseInProgress && (Time::millis() - _morseLastFrameTime > MORSE_FRAME_TIMEOUT_MS))
-        {
-            LOG_WARN(Logger::Nodes::J5Device, "Morse message timed out, discarding partial buffer");
-            resetMorseState();
-        }
-
         _driver.update();
-        _morsePlayer.update();
 
         if (_pbOpen.isClicked())
         {
@@ -80,7 +65,6 @@ class J5Device
             _driver.setCmd(0.0F);
         }
 
-        _solenoid.write(_morsePlayer.isActuatorOn() ? IO::eIOState::HIGH_ : IO::eIOState::LOW_);
 
         if (_timerCanSend.isReady())
         {
@@ -88,13 +72,13 @@ class J5Device
             armStatusMsg.data().currentPosition = 0.0F;  // No position feedback on joint yet
             armStatusMsg.data().currentSpeed = _driver.getCmd();
 
-            _jointCanDevice.sendMsg(armStatusMsg);
+            _canDevice.sendMsg(armStatusMsg);
         }
     }
 
     JointCanDeviceT& getUnderlyingCanDevice()
     {
-        return _jointCanDevice;
+        return _canDevice;
     }
 
   private:
@@ -104,94 +88,11 @@ class J5Device
         targetSpeed_ = cmd_.getData().targetSpeed;
     }
 
-    void CB_morseCodeStream(const RoverCan2::Msgs::MorseCode& msg_)
-    {
-        const auto& data = msg_.getData();
-        Serial.println("Here");
-
-        if (data.start)
-        {
-            // New message begins — reset regardless of any prior partial state
-            resetMorseState();
-
-            if (data.msg_length > MORSE_MAX_LEN)
-            {
-                LOG_ERROR(Logger::Nodes::J5Device,
-                          "Morse message length %u exceeds buffer of %u, dropping",
-                          static_cast<unsigned>(data.msg_length),
-                          static_cast<unsigned>(MORSE_MAX_LEN));
-                return;
-            }
-
-            _morseInProgress = true;
-            _morseExpectedLength = data.msg_length;
-        }
-
-        if (!_morseInProgress)
-        {
-            LOG_WARN(Logger::Nodes::J5Device, "Morse frame received with no message in progress, ignoring");
-            return;
-        }
-
-        if (data.index != _morseExpectedIndex)
-        {
-            LOG_WARN(Logger::Nodes::J5Device,
-                     "Morse frame loss: expected index %u, got %u",
-                     static_cast<unsigned>(_morseExpectedIndex),
-                     static_cast<unsigned>(data.index));
-            resetMorseState();
-            return;
-        }
-
-        _morseBuffer[data.index] = data.character;
-        _morseRunningChecksum = static_cast<uint8_t>(_morseRunningChecksum + data.character);
-        _morseLastFrameTime = Time::millis();
-
-        if (_morseRunningChecksum != data.checksum)
-        {
-            LOG_WARN(Logger::Nodes::J5Device,
-                     "Morse checksum mismatch at index %u, discarding message",
-                     static_cast<unsigned>(data.index));
-            resetMorseState();
-            return;
-        }
-
-        ++_morseExpectedIndex;
-
-        if (_morseExpectedIndex == _morseExpectedLength)
-        {
-            handleMorseComplete(_morseBuffer, _morseExpectedLength);
-            resetMorseState();
-        }
-    }
-
-    void handleMorseComplete(const uint8_t* buffer_, uint8_t length_)
-    {
-        // buffer_ is NOT null-terminated by itself; build a local terminated copy for logging/use
-        char text[MORSE_MAX_LEN + 1];
-        std::memcpy(text, buffer_, length_);
-        text[length_] = '\0';
-        Serial.printf("Morse message received (%u chars): %s\n", static_cast<unsigned>(length_), text);
-
-        _morsePlayer.start(buffer_, length_);
-    }
-
-    void resetMorseState()
-    {
-        _morseInProgress = false;
-        _morseExpectedIndex = 0;
-        _morseExpectedLength = 0;
-        _morseRunningChecksum = 0;
-    }
-
-    JointCanDeviceT _jointCanDevice
+    JointCanDeviceT _canDevice
         = JointCanDeviceT(RoverCan2::Constant::eDeviceId::GRIPPER_CLOSE_CONTROLLER,
                           RoverCan2::SubscriberMember<RoverCan2::Msgs::ArmJointCmd, J5Device>(*this, &J5Device::CB_canCmd),
                           RoverCan2::Publisher<RoverCan2::Msgs::ArmJointStatus>());
 
-    MorseCodeCanDeviceT _morseCanDevice = MorseCodeCanDeviceT(
-        RoverCan2::Constant::eDeviceId::MORSE_CODE,
-        RoverCan2::SubscriberMember<RoverCan2::Msgs::MorseCode, J5Device>(*this, &J5Device::CB_morseCodeStream));
 
     PWMGenerators::MCPWMTimer __pwmTimer = PWMGenerators::MCPWMTimer(1'000, PWMGenerators::MCPWMTimer::eMCPWMGroupID::GROUP_1);
     PWMGenerators::MCPWM __pwmGen = PWMGenerators::MCPWM(PIN_J5_PWM, __pwmTimer);
@@ -206,17 +107,6 @@ class J5Device
     float targetSpeed_ = 0.0F;
     LoopTimer<uint64_t, &Time::millis> _timerCanSend = {CAN_SEND_PERIOD_MS};
     Watchdog<uint64_t, &Time::millis> _canWatchdog = {CAN_WATCHDOG_VALIDITY_PERIOD};
-
-    // Morse code reassembly state
-    uint8_t _morseBuffer[MORSE_MAX_LEN] = {};
-    uint8_t _morseExpectedIndex = 0;
-    uint8_t _morseExpectedLength = 0;
-    uint8_t _morseRunningChecksum = 0;
-    bool _morseInProgress = false;
-    uint64_t _morseLastFrameTime = 0;
-
-    MorsePlayer _morsePlayer;
-    IO::DigitalOutput _solenoid = IO::DigitalOutput(PIN_USER_LED);
 
     // INA219 _currentSensor = INA219(Wire, 0x85, 0.05F, 4.0F);
 };
